chore(ts): conform with transformed sourceData expectations
